[refactor] Ajust shortdocs, docs and logo
