+++ lib/mix/tasks/phx_up.gen.plug.ex
@@ -1,8 +1,7 @@
 defmodule Mix.Tasks.PhxUp.Gen.Plug do
-  @shortdoc false
+  @shortdoc "Generate a plug file and a plug test file"
   @moduledoc false
 
-  alias PhoenixUp.Mix.Inflector
   use Mix.Task
 
   @doc false
@@ -48,6 +47,7 @@ defmodule Mix.Tasks.PhxUp.Gen.Plug do
     |> Mix.Generator.copy_template(final_path, opts)
   end
 
-  defp get_context([module|_]), do: Inflector.call(module)
+  defp get_context([module|_]), do: PhoenixUp.inflect(module)
   defp get_context([]), do: raise(RuntimeError, "Invalid module name")
 end
+
