+++ lib/mix/tasks/phx_up.gen.template.ex
@@ -1,8 +1,7 @@
 defmodule Mix.Tasks.PhxUp.Gen.Template do
-  @shortdoc false
+  @shortdoc "Generate a template file"
   @moduledoc false
 
-  alias PhoenixUp.Mix.Inflector
   use Mix.Task
 
   @doc false
@@ -39,6 +38,7 @@ defmodule Mix.Tasks.PhxUp.Gen.Template do
     |> Mix.Generator.copy_template(final_path, opts)
   end
 
-  defp get_context([module|_]), do: Inflector.call(module)
+  defp get_context([module|_]), do: PhoenixUp.inflect(module)
   defp get_context([]), do: raise(RuntimeError, "Invalid module name")
 end
+
